Experimenting ashmem IPC.
What's done in this commit:
TestApp has a Test Button. On click, it creates a MemoryFile object and get a file descriptor from the native. Then it sends the fd over AIDL IPC to universal device service, which passes it to native code and print in logcat.
Now, native code just print file descriptor.

Next step is to create shmem in universal device service's native. This link will be helpful: http://stackoverflow.com/questions/22436414/how-to-pass-file-descriptor-to-ashmem-between-processes